refactor(channels): 将 channel 和 secret 类型定义合并到 channels/types
